requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine learning libraries
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep learning (optional)
tensorflow>=2.13.0
keras>=2.13.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0

# Data ingestion
kaggle>=1.5.16
datasets>=2.14.0
huggingface-hub>=0.16.0

# Cloud storage (optional)
boto3>=1.26.0
google-cloud-storage>=2.8.0

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Model persistence
joblib>=1.3.0
pickle5>=0.0.12

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
